BACKEND/requirements.txt
openant
fastapi
uvicorn
pydantic
sqlalchemy
sqlalchemy-utils
psycopg2
pytest
python-dotenv
jwt
pyjwt
passlib
websockets